Move checkCheckable out of Infer.
Preparing for separate file with checkability logic.
